+++ src/components/FAQ.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const FAQ: React.FC = () => {
+    return (
+        <section id="faq" className="py-5">
+            <h2 className="text-center">Frequently Asked Questions</h2>
+            <p>Here you can find answers to the most common questions.</p>
+        </section>
+    );
+};
+
+export default FAQ;
