+++ src/components/Perks.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const Perks: React.FC = () => {
+    return (
+        <section id="perks" className="py-5">
+            <h2 className="text-center">Perks</h2>
+            <ul>
+                <li>Exclusive Offers</li>
+                <li>24/7 Support</li>
+                <li>Community Access</li>
+            </ul>
+        </section>
+    );
+};
+
+export default Perks;
